requirements.txt
ipython
Django==2.1.7
gunicorn==19.9.0
psycopg2-binary==2.7.7
djangorestframework
markdown
django-filter
django-allauth
django-rest-auth
model_mommy